+++ CHANGES.md
@@ -15,6 +15,8 @@ Version `1.0.0`:
 * *Breaking*: Renamed `@DiscardObjectAfterInvocation` to `@OnlyOnce`
 * *Breaking*: Renamed `@MayModify` to `@Modifies`. The latter has optional
   fields to describe when modification happens.
+* *Breaking*: Split `@Reentrancy` into `@Reentrant` and `@NotReentrant`
+
 
 Version `0.4`:
 * *Breaking*: Removed `LOW_LATENCY` and renamed `CACHE_BOUND` to `CACHE_HEAVY`